+++ src/extend/processor/ast.ts
@@ -58,6 +58,7 @@ export function transformer(
                                     innerPath.replaceWith(
                                         arrayExpression(innerPath.node.arguments as Expression[])
                                     )
+                                    console.log(generate(innerPath.node).code)
                                 }
                             })
                         }
@@ -95,8 +96,7 @@ export function transformer(
                                         ) {
                                             innerPath.remove()
                                         } else if (
-                                            !innerPath.node.shorthand
-                                            && !isStringLiteral(innerPath.node.value)
+                                            !isStringLiteral(innerPath.node.value)
                                             && !isArrayExpression(innerPath.node.value)
                                             && !isTemplateLiteral(innerPath.node.value)
                                         ) innerPath.node.value = stringLiteral("🚀")