+++ src/extend/processor/parser.ts
@@ -12,7 +12,7 @@ function reducer(args: ClassValue[], options?: TWGOptions) {
         if (!cur) return acc
         if (typeof cur === "object") {
             for (const [key, values] of Object.entries(cur)) {
-                (Array.isArray(values) ? values.flat(Infinity) : [values]).forEach(
+                (Array.isArray(values) ? values.flat(1 / 0) : [values]).forEach(
                     value => acc.push(
                         (parser(options)[key] as (...args: ClassValue[]) => string)(value as ClassValue)
                     )
@@ -22,7 +22,7 @@ function reducer(args: ClassValue[], options?: TWGOptions) {
             acc.push(...String(cur).split(" "))
         }
         return acc
-    }, []).flat().filter(Boolean)
+    }, []).flat().filter(value => value !== "")
 }
 
 /**